Fix macOS all-package build and localhost IPv4 dial

+++ src/internal/offalign/offalign_test.go
@@ -67,10 +67,10 @@ var alignCases = []struct {
 	before string
 	after  string
 	none   bool
-	blob   uint
-	base   uint
-	elem   uint
-	cnt    uint
+	blob   uintptr
+	base   uintptr
+	elem   uintptr
+	cnt    uintptr
 }{
 	{
 		name:   "no match",
